Extract main() from example projects
