+++ tasks/mpi/example/func_tests/main.cpp
@@ -225,14 +225,3 @@ TEST(Parallel_Operations_MPI, Test_Max_2) {
     ASSERT_EQ(reference_max[0], global_max[0]);
   }
 }
-
-int main(int argc, char** argv) {
-  boost::mpi::environment env(argc, argv);
-  boost::mpi::communicator world;
-  ::testing::InitGoogleTest(&argc, argv);
-  ::testing::TestEventListeners& listeners = ::testing::UnitTest::GetInstance()->listeners();
-  if (world.rank() != 0) {
-    delete listeners.Release(listeners.default_result_printer());
-  }
-  return RUN_ALL_TESTS();
-}
