+++ tasks/mpi/example/perf_tests/main.cpp
@@ -85,14 +85,3 @@ TEST(mpi_example_perf_test, test_task_run) {
     ASSERT_EQ(count_size_vector, global_sum[0]);
   }
 }
-
-int main(int argc, char** argv) {
-  boost::mpi::environment env(argc, argv);
-  boost::mpi::communicator world;
-  ::testing::InitGoogleTest(&argc, argv);
-  ::testing::TestEventListeners& listeners = ::testing::UnitTest::GetInstance()->listeners();
-  if (world.rank() != 0) {
-    delete listeners.Release(listeners.default_result_printer());
-  }
-  return RUN_ALL_TESTS();
-}
