+++ tasks/omp/example/func_tests/main.cpp
@@ -194,8 +194,3 @@ TEST(Parallel_Operations_OpenMP, Test_Mult_2) {
   testOmpTaskParallel.post_processing();
   ASSERT_EQ(ref_res[0], par_res[0]);
 }
-
-int main(int argc, char **argv) {
-  ::testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
