+++ tasks/seq/example/func_tests/main.cpp
@@ -118,8 +118,3 @@ TEST(Sequential, Test_Sum_100) {
   testTaskSequential.post_processing();
   ASSERT_EQ(count, out[0]);
 }
-
-int main(int argc, char **argv) {
-  testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
