+++ tasks/seq/example/perf_tests/main.cpp
@@ -78,8 +78,3 @@ TEST(sequential_example_perf_test, test_task_run) {
   ppc::core::Perf::print_perf_statistic(perfResults);
   ASSERT_EQ(count, out[0]);
 }
-
-int main(int argc, char **argv) {
-  testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
