+++ tasks/tbb/example/perf_tests/main.cpp
@@ -71,8 +71,3 @@ TEST(tbb_example_perf_test, test_task_run) {
   ppc::core::Perf::print_perf_statistic(perfResults);
   ASSERT_EQ(count + 1, out[0]);
 }
-
-int main(int argc, char **argv) {
-  testing::InitGoogleTest(&argc, argv);
-  return RUN_ALL_TESTS();
-}
